ui/requirements.txt
streamlit>=1.32
requests>=2.31
pillow>=10.2

